Rename ATP responders
