+++ responders/Zscaler/ZscalerZIA_ATPSecurityExceptions.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
 # encoding: utf-8
 """
-Zscaler ZIA Blacklist Responder
+Zscaler ZIA ATP Security Exceptions Responder
 
 Adds or removes URLs (domain, fqdn, url) to/from Zscaler Internet Access
-global URL blacklist.
+ATP Security Exceptions (Policy > Security > Advanced Threat Protection > Security Exceptions).
 """
 
 import re
@@ -12,15 +12,14 @@ import time
 from datetime import datetime
 from urllib.parse import urlparse
 
-import tldextract
 from cortexutils.responder import Responder
 from zscaler import ZscalerClient
 from zscaler.oneapi_client import LegacyZIAClient
 
 
-class ZscalerZIA_Blacklist(Responder):
+class ZscalerZIA_ATPSecurityExceptions(Responder):
     """
-    Cortex Responder to manage URL blacklist in Zscaler Internet Access (ZIA)
+    Cortex Responder to manage ATP Security Exceptions in Zscaler Internet Access (ZIA)
     Supports: domain, fqdn, url
     """
 
@@ -30,39 +29,50 @@ class ZscalerZIA_Blacklist(Responder):
         'io', 'co', 'uk', 'de', 'fr', 'jp', 'cn', 'au'
     }
 
-
     def __init__(self):
         Responder.__init__(self)
 
+        # Determine authentication type
         self.auth_type = self.get_param('config.auth_type', 'oneapi').lower()
 
+        # Validate auth_type
         if self.auth_type not in ['oneapi', 'legacy']:
             self.error(f"Invalid auth_type: {self.auth_type}. Must be 'oneapi' or 'legacy'")
 
+        # Authentication parameters - conditionally required based on auth_type
         if self.auth_type == 'oneapi':
+            # OneAPI OAuth2 parameters
             self.zia_vanity_domain = self.get_param('config.zia_vanity_domain', None, 'ZIA Vanity Domain is required for OneAPI authentication')
             self.zia_client_id = self.get_param('config.zia_client_id', None, 'ZIA Client ID is required for OneAPI authentication')
             self.zia_client_secret = self.get_param('config.zia_client_secret', None, 'ZIA Client Secret is required for OneAPI authentication')
-            self.zia_cloud = self.get_param('config.zia_cloud', None)
+            # Legacy parameters not used for OneAPI
             self.zia_username = None
             self.zia_password = None
             self.zia_api_key = None
-        else:
+        else:  # legacy
+            # Legacy API parameters
             self.zia_username = self.get_param('config.zia_username', None, 'ZIA Username is required for legacy authentication')
             self.zia_password = self.get_param('config.zia_password', None, 'ZIA Password is required for legacy authentication')
             self.zia_api_key = self.get_param('config.zia_api_key', None, 'ZIA API Key is required for legacy authentication')
             self.zia_cloud = self.get_param('config.zia_cloud', None, 'ZIA Cloud is required for legacy authentication')
+            # OneAPI parameters not used for legacy
             self.zia_vanity_domain = None
             self.zia_client_id = None
             self.zia_client_secret = None
+
+        # Optional config
+        if self.auth_type == 'oneapi':
+            self.zia_cloud = self.get_param('config.zia_cloud', None)  # Optional: for beta/alpha environments
 
         self.action_type = self.get_param('config.action_type', 'add').lower()  # 'add' or 'remove' - set by responder flavor JSON
         self.dry_run = self.get_param('config.dry_run', False)
         self.activate_changes = self.get_param('config.activate_changes', True)
         self.allow_risky_iocs = self.get_param('config.allow_risky_iocs', False)
         self.allow_wildcards = self.get_param('config.allow_wildcards', False)
-        self.http_proxy_hostname = self.get_param('config.http_proxy_hostname', None)
-        self.http_proxy_port = self.get_param('config.http_proxy_port', None)
+        proxy_url = self.get_param('config.proxy_https', None) or self.get_param('config.proxy_http', None)
+        self.proxy_config = self._parse_proxy(proxy_url)
+
+        # Initialize ZIA client
         self.zia_client = None
         self.audit_data = {
             'timestamp': datetime.utcnow().isoformat() + 'Z',
@@ -74,22 +84,36 @@ class ZscalerZIA_Blacklist(Responder):
             'errors': []
         }
 
+    @staticmethod
+    def _parse_proxy(proxy_url):
+        if not proxy_url:
+            return None
+        parsed = urlparse(proxy_url)
+        cfg = {"host": parsed.hostname}
+        if parsed.port:
+            cfg["port"] = parsed.port
+        return cfg
+
     def _init_zia_client(self):
+        """Initialize Zscaler ZIA client with OneAPI OAuth2 or Legacy authentication"""
         if self.zia_client is None:
             try:
                 if self.auth_type == 'oneapi':
+                    # Prepare OneAPI OAuth2 configuration
                     config = {
                         "clientId": self.zia_client_id,
                         "clientSecret": self.zia_client_secret,
                         "vanityDomain": self.zia_vanity_domain,
                         "logging": {"enabled": False, "verbose": False}
                     }
+
+                    # Add cloud parameter if specified (for beta/alpha environments)
                     if self.zia_cloud:
                         config["cloud"] = self.zia_cloud
+                    if self.proxy_config:
+                        config["proxy"] = self.proxy_config
 
-                    if self.http_proxy_hostname and self.http_proxy_port:
-                        config["proxy"] = {"host": self.http_proxy_hostname, "port": self.http_proxy_port}
-
+                    # Create ZscalerClient and access ZIA services via .zia
                     try:
                         zscaler_client = ZscalerClient(config)
                         self.zia_client = zscaler_client.zia
@@ -97,7 +121,9 @@ class ZscalerZIA_Blacklist(Responder):
                         error_msg = f"OAuth authentication failed: {str(oauth_error)}"
                         self.audit_data['errors'].append(error_msg)
                         self.error(error_msg)
-                else:
+
+                else:  # legacy
+                    # Prepare Legacy API configuration
                     config = {
                         "username": self.zia_username,
                         "password": self.zia_password,
@@ -106,9 +132,10 @@ class ZscalerZIA_Blacklist(Responder):
                         "logging": {"enabled": False, "verbose": False}
                     }
 
-                    if self.http_proxy_hostname and self.http_proxy_port:
-                        config["proxy"] = {"host": self.http_proxy_hostname, "port": self.http_proxy_port}
+                    if self.proxy_config:
+                        config["proxy"] = self.proxy_config
 
+                    # Create LegacyZIAClient
                     try:
                         legacy_client = LegacyZIAClient(config)
                         self.zia_client = legacy_client.zia
@@ -118,40 +145,40 @@ class ZscalerZIA_Blacklist(Responder):
                         self.error(error_msg)
 
             except Exception as e:
+                # Catch any other unexpected errors
                 error_msg = f'Unexpected error during ZIA client initialization: {str(e)}'
                 self.audit_data['errors'].append(error_msg)
                 self.error(error_msg)
 
     def _validate_domain(self, domain):
+        """Validate and normalize domain/FQDN"""
         if not domain:
             return False, None, "Empty domain"
 
+        # Remove protocol if present
         domain = domain.lower().strip()
         domain = re.sub(r'^https?://', '', domain)
-        domain = domain.split('/')[0]
+        domain = domain.split('/')[0]  # Remove path
 
+        # Check for wildcard
         if domain.startswith('*.'):
             if not self.allow_wildcards:
                 return False, None, f"Wildcard domains not allowed: {domain}"
             domain = '.' + domain[2:]  # Zscaler wildcard format: .example.com
             domain_to_check = domain[1:]
-        elif self.allow_wildcards:
-            # Use Mozilla Public Suffix List to find the registrable domain.
-            # sub.evil.co.uk -> .evil.co.uk, xyz.app.evil.com -> .evil.com
-            extracted = tldextract.extract(domain)
-            parent = extracted.registered_domain or domain
-            domain = '.' + parent
-            domain_to_check = parent
         else:
             domain_to_check = domain
 
+        # Check for bare TLD
         if not self.allow_risky_iocs and domain_to_check in self.BARE_TLDS:
             return False, None, f"Bare TLD blocked (too generic): {domain}"
 
+        # Basic FQDN validation
         fqdn_pattern = r'^(?:[a-zA-Z0-9](?:[a-zA-Z0-9\-]{0,61}[a-zA-Z0-9])?\.)*[a-zA-Z0-9](?:[a-zA-Z0-9\-]{0,61}[a-zA-Z0-9])?$'
         if not re.match(fqdn_pattern, domain_to_check):
             return False, None, f"Invalid domain format: {domain}"
 
+        # Check for overly generic patterns
         if not self.allow_risky_iocs:
             parts = domain_to_check.split('.')
             if len(parts) <= 1:
@@ -160,10 +187,12 @@ class ZscalerZIA_Blacklist(Responder):
         return True, domain, None
 
     def _validate_url(self, url):
+        """Validate URL and return full URL without protocol"""
         if not url:
             return False, None, None, "Empty URL"
 
         try:
+            # Ensure URL has a scheme for parsing
             if not url.startswith(('http://', 'https://')):
                 url = 'http://' + url
 
@@ -173,11 +202,13 @@ class ZscalerZIA_Blacklist(Responder):
             if not hostname:
                 return False, None, None, "Could not extract hostname from URL"
 
+            # Validate the hostname part
             is_valid, normalized_host, error = self._validate_domain(hostname)
             if not is_valid:
                 return False, None, None, error
 
-            full_url_no_protocol = hostname
+            # Build full URL without protocol: hostname + path + params + query
+            full_url_no_protocol = normalized_host
             if parsed.path:
                 full_url_no_protocol += parsed.path
             if parsed.params:
@@ -190,6 +221,7 @@ class ZscalerZIA_Blacklist(Responder):
             return False, None, None, f"URL parsing error: {str(e)}"
 
     def _activate_changes(self):
+        """Activate pending changes in Zscaler"""
         if self.dry_run or not self.activate_changes:
             return True, "Activation skipped"
 
@@ -212,37 +244,37 @@ class ZscalerZIA_Blacklist(Responder):
                 return False, f"Activation error: {str(e)}"
         return False, "Activation failed after retries"
 
-    def _manage_blacklist(self, urls):
+    def _manage_allowlist(self, urls):
         """
-        Add or remove URLs from ZIA global blacklist using incremental updates
+        Add or remove URLs from ZIA global allowlist using incremental updates
         Returns: (success, already_present, error_message)
         """
         self._init_zia_client()
 
         try:
-            # Get current blacklist - SDK returns (data, response, error)
-            blacklist_data, _, err = self.zia_client.security_policy_settings.get_blacklist()
+            # Get current allowlist (security settings) - SDK returns (data, response, error)
+            security_settings, _, err = self.zia_client.security_policy_settings.get_whitelist()
 
             if err:
-                return False, False, f"Failed to get blacklist: {err}"
+                return False, False, f"Failed to get allowlist: {err}"
 
-            # Extract current blacklisted URLs (normalize to lowercase for comparison)
-            # Try different possible field names (SDK uses snake_case: blacklist_urls)
-            if hasattr(blacklist_data, 'blacklist_urls'):
-                blacklist_urls = blacklist_data.blacklist_urls or []
-            elif hasattr(blacklist_data, 'blacklistUrls'):
-                blacklist_urls = blacklist_data.blacklistUrls or []
-            elif isinstance(blacklist_data, dict):
-                blacklist_urls = blacklist_data.get('blacklistUrls', blacklist_data.get('blacklist_urls', []))
+            # Extract current allowlisted URLs (normalize to lowercase for comparison)
+            # Try different possible field names (SDK uses snake_case: whitelist_urls)
+            if hasattr(security_settings, 'whitelist_urls'):
+                allowlist_urls = security_settings.whitelist_urls or []
+            elif hasattr(security_settings, 'whitelistUrls'):
+                allowlist_urls = security_settings.whitelistUrls or []
+            elif isinstance(security_settings, dict):
+                allowlist_urls = security_settings.get('whitelistUrls', security_settings.get('whitelist_urls', []))
             else:
-                blacklist_urls = []
+                allowlist_urls = []
 
-            current_urls = set(url.lower() for url in blacklist_urls if url)
+            current_urls = set(url.lower() for url in allowlist_urls if url)
 
-            # Debug: Log blacklist info (useful for troubleshooting)
-            self.audit_data['blacklist_info'] = {
+            # Debug: Log allowlist info (useful for troubleshooting)
+            self.audit_data['allowlist_info'] = {
                 'total_urls': len(current_urls),
-                'response_type': type(blacklist_data).__name__
+                'response_type': type(security_settings).__name__
             }
 
             # Normalize URLs to add/remove
@@ -253,17 +285,30 @@ class ZscalerZIA_Blacklist(Responder):
             for url in urls:
                 urls_to_process.add(url.lower())
 
+            # Debug logging - useful for troubleshooting
+            self.audit_data['debug'] = {
+                'url_to_process': list(urls_to_process)[0] if urls_to_process else None,
+                'found_in_allowlist': urls_to_process.issubset(current_urls),
+                'similar_urls': [
+                    al_url for al_url in current_urls
+                    for proc_url in urls_to_process
+                    if proc_url.split('/')[0] in al_url or al_url.split('/')[0] in proc_url
+                ][:5]
+            }
+
             # Check if action is needed
-            urls_exist_in_blacklist = urls_to_process.issubset(current_urls)
+            urls_exist_in_allowlist = urls_to_process.issubset(current_urls)
 
             # For add: skip if already present
             # For remove: skip if NOT present
             if self.action_type == 'add':
-                if urls_exist_in_blacklist:
+                if urls_exist_in_allowlist:
+                    # Already allowlisted
                     self.audit_data['already_present'] = True
                     return True, True, None
             else:  # remove
-                if not urls_exist_in_blacklist:
+                if not urls_exist_in_allowlist:
+                    # Already not in allowlist (not found)
                     self.audit_data['already_present'] = True
                     return True, True, None
 
@@ -273,12 +318,12 @@ class ZscalerZIA_Blacklist(Responder):
 
                 # Use appropriate method based on action type
                 if self.action_type == 'add':
-                    _, _, err = self.zia_client.security_policy_settings.add_urls_to_blacklist(urls_list)
+                    _, _, err = self.zia_client.security_policy_settings.add_urls_to_whitelist(urls_list)
                 else:  # remove
-                    _, _, err = self.zia_client.security_policy_settings.delete_urls_from_blacklist(urls_list)
+                    _, _, err = self.zia_client.security_policy_settings.delete_urls_from_whitelist(urls_list)
 
                 if err:
-                    return False, False, f"Failed to update blacklist: {err}"
+                    return False, False, f"Failed to update allowlist: {err}"
 
                 # Activate changes if enabled
                 activation_success, activation_msg = self._activate_changes()
@@ -313,8 +358,8 @@ class ZscalerZIA_Blacklist(Responder):
                     self.audit_data['errors'].append(error)
                     self.error(error)
 
-                self.audit_data['action_taken'] = 'manage_blacklist'
-                success, already_present, error = self._manage_blacklist(normalized)
+                self.audit_data['action_taken'] = 'manage_allowlist'
+                success, already_present, error = self._manage_allowlist(normalized)
 
                 if not success:
                     self.audit_data['errors'].append(error)
@@ -330,22 +375,22 @@ class ZscalerZIA_Blacklist(Responder):
                 else:
                     status = action_verb
 
-                message = f"Domain '{normalized}' {status} ZIA URL blacklist"
+                message = f"Domain '{normalized}' {status} ZIA URL allowlist"
 
                 if self.dry_run and not already_present:
-                    message = f"[DRY RUN] Would have {action_verb} ZIA URL blacklist"
+                    message = f"[DRY RUN] Would have {action_verb} ZIA URL allowlist"
 
                 self.report({'message': message, 'audit': self.audit_data})
 
             elif data_type == 'url':
-                is_valid, normalized_url, full_url, error = self._validate_url(data_value)
+                is_valid, _normalized_url, full_url, error = self._validate_url(data_value)
                 if not is_valid:
                     self.audit_data['errors'].append(error)
                     self.error(error)
 
-                self.audit_data['action_taken'] = 'manage_blacklist'
-                # For URL blacklisting, we add the full URL (without protocol)
-                success, already_present, error = self._manage_blacklist(full_url)
+                self.audit_data['action_taken'] = 'manage_allowlist'
+                # For URL allowlisting, we add the full URL (without protocol)
+                success, already_present, error = self._manage_allowlist(full_url)
 
                 if not success:
                     self.audit_data['errors'].append(error)
@@ -361,10 +406,10 @@ class ZscalerZIA_Blacklist(Responder):
                 else:
                     status = action_verb
 
-                message = f"URL '{data_value}' (added as: '{full_url}') {status} ZIA URL blacklist"
+                message = f"URL '{data_value}' (added as: '{full_url}') {status} ZIA URL allowlist"
 
                 if self.dry_run and not already_present:
-                    message = f"[DRY RUN] Would have {action_verb} ZIA URL blacklist"
+                    message = f"[DRY RUN] Would have {action_verb} ZIA URL allowlist"
 
                 self.report({'message': message, 'audit': self.audit_data})
 
@@ -379,22 +424,12 @@ class ZscalerZIA_Blacklist(Responder):
             self.error(error_msg)
 
     def operations(self, raw):
-        """Define operations to perform on TheHive after responder execution"""
-        # Get timestamp for tag (format: YYYY-MM-DD_HH:MM:SS)
-        timestamp = datetime.utcnow().strftime('%Y-%m-%d_%H:%M:%S')
-
         if self.audit_data.get('already_present'):
-            if self.action_type == 'add':
-                tag = f'ZIA:blacklist:already-present:{timestamp}'
-            else:
-                tag = f'ZIA:blacklist:not-found:{timestamp}'
+            tag = 'ZIA:atp-security-exceptions:already-present' if self.action_type == 'add' else 'ZIA:atp-security-exceptions:not-found'
         elif self.dry_run:
-            tag = f'ZIA:blacklist:dry-run:{timestamp}'
+            tag = 'ZIA:atp-security-exceptions:dry-run'
         else:
-            if self.action_type == 'add':
-                tag = f'ZIA:blacklist:added:{timestamp}'
-            else:
-                tag = f'ZIA:blacklist:removed:{timestamp}'
+            tag = 'ZIA:atp-security-exceptions:added' if self.action_type == 'add' else 'ZIA:atp-security-exceptions:removed'
 
         return [
             self.build_operation('AddTagToCase', tag='ZIA:action-taken'),
@@ -403,4 +438,4 @@ class ZscalerZIA_Blacklist(Responder):
 
 
 if __name__ == '__main__':
-    ZscalerZIA_Blacklist().run()
+    ZscalerZIA_ATPSecurityExceptions().run()
