+++ responders/Zscaler/ZscalerZIA_AddToATPDenylistWildcard.json
@@ -1,12 +1,14 @@
 {
-  "name": "ZscalerZIA_AddToBlacklistWildcard",
+  "name": "ZscalerZIA_AddToATPDenylistWildcard",
   "version": "1.0",
   "author": "Fabien Bloume, StrangeBee",
   "url": "https://github.com/TheHive-Project/Cortex-Analyzers",
   "license": "AGPL-V3",
-  "description": "Add a domain and all its subdomains (*.example.com) to Zscaler Internet Access Blocked Malicious URLs - Denylist specific URLs (Policy > Security > Advanced Threat Protection > Blocked Malicious URLs)",
-  "dataTypeList": ["thehive:case_artifact"],
-  "command": "Zscaler/ZscalerZIA_Blacklist.py",
+  "description": "Block a domain and all its subdomains in the ZIA ATP Denylist (Policy > Security > Advanced Threat Protection > Blocked Malicious URLs)",
+  "dataTypeList": [
+    "thehive:case_artifact"
+  ],
+  "command": "Zscaler/ZscalerZIA_ATPDenylist.py",
   "baseConfig": "ZscalerZIA",
   "config": {
     "action_type": "add",
@@ -23,7 +25,7 @@
     },
     {
       "name": "zia_vanity_domain",
-      "description": "[OneAPI only] Zscaler ZIdentity vanity domain for your organization (e.g., 'acme' from acme.zslogin.net)",
+      "description": "[OneAPI only] Zscaler ZIdentity vanity domain for your organization (example: 'acme' from acme.zslogin.net)",
       "type": "string",
       "multi": false,
       "required": false
@@ -65,7 +67,7 @@
     },
     {
       "name": "zia_cloud",
-      "description": "Cloud environment name. Required for legacy auth (e.g., 'zscaler', 'zscalerone', 'zscalertwo'). Optional for OneAPI (use for beta/alpha environments).",
+      "description": "Cloud environment name. Required for legacy auth (zscaler, zscalerone, zscalertwo). Optional for OneAPI (use for beta/alpha environments).",
       "type": "string",
       "multi": false,
       "required": false
@@ -88,25 +90,11 @@
     },
     {
       "name": "allow_risky_iocs",
-      "description": "Allow adding risky IOCs like bare TLDs (e.g., 'com') to blacklist",
+      "description": "Allow adding risky IOCs like bare TLDs (like 'com') to blacklist",
       "type": "boolean",
       "multi": false,
       "required": false,
       "defaultValue": false
-    },
-    {
-      "name": "http_proxy_hostname",
-      "description": "HTTP proxy hostname if Cortex requires one to reach the Zscaler API (e.g., 'proxy.internal')",
-      "type": "string",
-      "multi": false,
-      "required": false
-    },
-    {
-      "name": "http_proxy_port",
-      "description": "HTTP proxy port (e.g., 3128)",
-      "type": "number",
-      "multi": false,
-      "required": false
     }
   ],
   "registration_required": true,
