+++ responders/Zscaler/ZscalerZIA_RemoveFromATPDenylistWildcard.json
@@ -1,15 +1,18 @@
 {
-  "name": "ZscalerZIA_RemoveFromSecurityExceptions",
+  "name": "ZscalerZIA_RemoveFromATPDenylistWildcard",
   "version": "1.0",
   "author": "Fabien Bloume, StrangeBee",
   "url": "https://github.com/TheHive-Project/Cortex-Analyzers",
   "license": "AGPL-V3",
-  "description": "Remove URLs (domain, fqdn, url) from Zscaler Internet Access Security Exceptions - Do Not Scan Content from these URLs (Policy > Security > Advanced Threat Protection > Security Exceptions)",
-  "dataTypeList": ["thehive:case_artifact"],
-  "command": "Zscaler/ZscalerZIA_Allowlist.py",
+  "description": "Remove a domain and all its subdomains from the ZIA ATP Denylist (Policy > Security > Advanced Threat Protection > Blocked Malicious URLs)",
+  "dataTypeList": [
+    "thehive:case_artifact"
+  ],
+  "command": "Zscaler/ZscalerZIA_ATPDenylist.py",
   "baseConfig": "ZscalerZIA",
   "config": {
-    "action_type": "remove"
+    "action_type": "remove",
+    "allow_wildcards": true
   },
   "configurationItems": [
     {
@@ -22,7 +25,7 @@
     },
     {
       "name": "zia_vanity_domain",
-      "description": "[OneAPI only] Zscaler ZIdentity vanity domain for your organization (e.g., 'acme' from acme.zslogin.net)",
+      "description": "[OneAPI only] Zscaler ZIdentity vanity domain for your organization (example: 'acme' from acme.zslogin.net)",
       "type": "string",
       "multi": false,
       "required": false
@@ -64,7 +67,7 @@
     },
     {
       "name": "zia_cloud",
-      "description": "Cloud environment name. Required for legacy auth (e.g., 'zscaler', 'zscalerone', 'zscalertwo'). Optional for OneAPI (use for beta/alpha environments).",
+      "description": "Cloud environment name. Required for legacy auth (zscaler, zscalerone, zscalertwo). Optional for OneAPI (use for beta/alpha environments).",
       "type": "string",
       "multi": false,
       "required": false
@@ -87,33 +90,11 @@
     },
     {
       "name": "allow_risky_iocs",
-      "description": "Allow removing risky IOCs like bare TLDs (e.g., 'com') from allowlist",
+      "description": "Allow removing risky IOCs like bare TLDs (like 'com') from blacklist",
       "type": "boolean",
       "multi": false,
       "required": false,
       "defaultValue": false
-    },
-    {
-      "name": "allow_wildcards",
-      "description": "Allow wildcard domains (e.g., *.example.com)",
-      "type": "boolean",
-      "multi": false,
-      "required": false,
-      "defaultValue": false
-    },
-    {
-      "name": "http_proxy_hostname",
-      "description": "HTTP proxy hostname if Cortex requires one to reach the Zscaler API (e.g., 'proxy.internal')",
-      "type": "string",
-      "multi": false,
-      "required": false
-    },
-    {
-      "name": "http_proxy_port",
-      "description": "HTTP proxy port (e.g., 3128)",
-      "type": "number",
-      "multi": false,
-      "required": false
     }
   ],
   "registration_required": true,
