+++ responders/Zscaler/ZscalerZIA_RemoveFromATPSecurityExceptions.json
@@ -1,15 +1,17 @@
 {
-  "name": "ZscalerZIA_AddToBlacklist",
+  "name": "ZscalerZIA_RemoveFromATPSecurityExceptions",
   "version": "1.0",
   "author": "Fabien Bloume, StrangeBee",
   "url": "https://github.com/TheHive-Project/Cortex-Analyzers",
   "license": "AGPL-V3",
-  "description": "Add URLs (domain, fqdn, url) to Zscaler Internet Access Blocked Malicious URLs - Denylist specific URLs (Policy > Security > Advanced Threat Protection > Blocked Malicious URLs)",
-  "dataTypeList": ["thehive:case_artifact"],
-  "command": "Zscaler/ZscalerZIA_Blacklist.py",
+  "description": "Remove domains, FQDNs and URLs from ZIA Security Exceptions (Policy > Security > Advanced Threat Protection > Security Exceptions)",
+  "dataTypeList": [
+    "thehive:case_artifact"
+  ],
+  "command": "Zscaler/ZscalerZIA_ATPSecurityExceptions.py",
   "baseConfig": "ZscalerZIA",
   "config": {
-    "action_type": "add"
+    "action_type": "remove"
   },
   "configurationItems": [
     {
@@ -22,7 +24,7 @@
     },
     {
       "name": "zia_vanity_domain",
-      "description": "[OneAPI only] Zscaler ZIdentity vanity domain for your organization (e.g., 'acme' from acme.zslogin.net)",
+      "description": "[OneAPI only] Zscaler ZIdentity vanity domain for your organization (example: 'acme' from acme.zslogin.net)",
       "type": "string",
       "multi": false,
       "required": false
@@ -64,7 +66,7 @@
     },
     {
       "name": "zia_cloud",
-      "description": "Cloud environment name. Required for legacy auth (e.g., 'zscaler', 'zscalerone', 'zscalertwo'). Optional for OneAPI (use for beta/alpha environments).",
+      "description": "Cloud environment name. Required for legacy auth (zscaler, zscalerone, zscalertwo). Optional for OneAPI (use for beta/alpha environments).",
       "type": "string",
       "multi": false,
       "required": false
@@ -87,7 +89,7 @@
     },
     {
       "name": "allow_risky_iocs",
-      "description": "Allow adding risky IOCs like bare TLDs (e.g., 'com') to blacklist",
+      "description": "Allow removing risky IOCs like bare TLDs (like 'com') from allowlist",
       "type": "boolean",
       "multi": false,
       "required": false,
@@ -95,25 +97,11 @@
     },
     {
       "name": "allow_wildcards",
-      "description": "Allow wildcard domains (e.g., *.example.com)",
+      "description": "Allow wildcard domains (like *.example.com)",
       "type": "boolean",
       "multi": false,
       "required": false,
       "defaultValue": false
-    },
-    {
-      "name": "http_proxy_hostname",
-      "description": "HTTP proxy hostname if Cortex requires one to reach the Zscaler API (e.g., 'proxy.internal')",
-      "type": "string",
-      "multi": false,
-      "required": false
-    },
-    {
-      "name": "http_proxy_port",
-      "description": "HTTP proxy port (e.g., 3128)",
-      "type": "number",
-      "multi": false,
-      "required": false
     }
   ],
   "registration_required": true,
